pyproject.toml
[project]
name = "cloud-stock-analytics-databricks"
version = "1.0.0"
description = "Microservices-style stock analytics and prediction platform for Databricks and Delta Lake."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2.0",
    "pyarrow>=16.0.0",
    "PyYAML>=6.0.1",
    "yfinance>=0.2.54",
    "pyspark>=3.5.0",
    "delta-spark>=3.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "plotly>=5.24.0",
    "dash>=2.18.0",
]

[tool.setuptools]
py-modules = []

